Use subset of configurable attributes in configurable products
Allows for a configurable product to be configurable on a subset of the
attributes that can be used. For example if an attribute set has three
attributes that can be used to configure a configurable product, but the
admin only wants to use two, they can be set in the productData array when
calling product.create.

array (
  'configurable_attributes' =>
  array (
    0 => 'size',
    1 => 'color',
  ),
)